+++ KinectSDKandOF/slider.h
@@ -9,14 +9,14 @@ public:
 	//data
 	int originX,originY;
 	int sWidth;
-	int value;
+	float value;
 	int sliderPosX,sliderPosY;
 	ofTrueTypeFont font;
 	ofImage img;
 
 	slider(int x, int y, int sWidth);
 
-	void drawSlider(int startFrom, int endTo);
+	void drawSlider(float startFrom, float endTo);
 	void getSliderPosX(int _mouseX, int _mouseY);
 };
